Move mask dilation from predict to integrate
Previously laue.predict dilated the detector mask by a radius estimated from
the full predicted set, while laue.integrate estimated its integration-window
radius from the mask-filtered set. The two radii were derived from different
spot populations and could diverge, so an integration window could reach into
masked pixels the dilation was sized to exclude.

Consolidate the radius-dependent masking in laue.integrate, which now estimates
a single radius from the predicted set and reuses it for both the mask dilation
and the integration window. Dilation radius == window radius by construction, so
no surviving centroid's window can overlap a masked pixel.

laue.predict now performs only the radius-independent step: dropping predictions
whose centroid lands directly on a masked pixel. The shared radius/mask helpers
(estimate_integration_radius, unmasked_prediction_selection) move to
laue_dials.algorithms.integration.

Regenerate the regression fixtures on images 15/16 (both refine below half a
pixel: 0.421/0.436 px) with the new pipeline; baseline Pearson CC vs
Precognition is 0.977 over ~2290 matched reflections. Add regenerate_fixtures.sh
documenting the full generation sequence.
