+++ tests/regression/test_integration_regression.py
@@ -12,12 +12,13 @@ detector mask (gzipped) and the Precognition reference table live in the
 repository. If SBGrid cannot be reached the test is marked xfail rather than
 failing outright, per the review discussion on PR #86.
 
-The predicted reflection table was generated by ``laue.predict``, which uses the
-detector mask (``pixels.mask``) to discard predictions landing in masked regions
-of the detector. The mask is therefore shipped alongside the fixtures so the
-prediction step is reproducible; ``laue.integrate`` itself does not read it, so
-it does not affect the correlation computed here. See ``tests/data/regression/
-README.md`` for the full provenance of every fixture.
+The predicted reflection table was generated by ``laue.predict``, which drops
+predictions landing directly on masked detector pixels. ``laue.integrate`` then
+reads the same detector mask (``pixels.mask``), dilates it by the integration
+radius, and discards predictions whose integration window would overlap it, so
+the mask is shipped alongside the fixtures and is applied when this test runs.
+See ``tests/data/regression/README.md`` for the full provenance of every
+fixture.
 """
 
 import gzip
@@ -30,10 +31,12 @@ import numpy as np
 import pytest
 
 # SBGrid dataset 1117 (DHFR Laue), served from the dataset root over rsync.
+# Images 15 and 16 were chosen because both have laue.compute_rmsds RMSDs below
+# half a pixel (0.421 / 0.436 px) and are adjacent.
 SBGRID_RSYNC_ROOT = "rsync://data.sbgrid.org/10.15785/SBGRID/1117"
-IMAGES = ("e080_001.mccd", "e080_002.mccd")
+IMAGES = ("e080_015.mccd", "e080_016.mccd")
 
-# Baseline established on this branch is Pearson CC ~= 0.978 over ~2200 matched
+# Baseline established on this branch is Pearson CC ~= 0.977 over ~2290 matched
 # reflections. Thresholds are set below the baseline with margin for numerical
 # variation, so only a genuine regression trips them.
 MIN_PEARSON_CC = 0.95
@@ -79,9 +82,9 @@ def _relocate_experiments(image_dir, out_path):
     mask (written next to ``out_path``), then write it to ``out_path``.
 
     The mask is shipped gzipped and expanded here because dxtbx expects an
-    uncompressed pickle. ``laue.integrate`` does not consult the mask, but
-    pointing at a real file keeps the relocated experiment self-consistent and
-    reproducible by ``laue.predict``.
+    uncompressed pickle. ``laue.integrate`` dilates this mask by the integration
+    radius and drops predictions whose window would overlap it, so pointing the
+    relocated experiment at the real mask file is required for the test.
     """
     mask_path = os.path.join(os.path.dirname(out_path), "pixels.mask")
     with gzip.open(MASK_FIXTURE_GZ, "rb") as src, open(mask_path, "wb") as dst:
